refactor(cli): extract launcher backends into axolotl.cli.launchers package
Mechanical move of build_command/_launch_* helpers out of cli/utils/train.py;
back-compat re-exports stay in place. No behavior change.
